Se terminó la Pantalla Tipo
