+++ src/styles/views/_vistaOfertas.scss
@@ -2,18 +2,15 @@ body{
   background-color: #E9ECF5;
 }
 
-.container-titulo-ofertas{
+.title{
+  font-size: $font-size-xl;
+  font-family: Arial, Helvetica, sans-serif;
 }
-.ofertas-categoria {
-  border: 0px solid orangered;
+.container-titulo-ofertas{
   height: 15vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
 }
 
 .ofertas-imagenes {
-  border: 5px solid orangered;
   height: 70vh;
   display: flex;
   justify-content: center;
@@ -22,7 +19,10 @@ body{
 }
 
 .ofertas-ropa-ordenada{
+  height: 90%;
+  width: 90%;
   display: flex;
+  flex-wrap: wrap;
   justify-content: center;
   align-items: center;
 }
