--- src/views/Ofertas.js
@@ -1,127 +1,0 @@
-import { Header } from "../components/Header"
-import { Card } from "../components/Card.js"
-import { Soporte } from "../components/Soporte"
-
-
-const iconcitoLink = 'https://res.cloudinary.com/initial-cloud/image/upload/v1620870788/carrito_zs5b0x.png'
-const productos = [
-  {
-    id: 1,
-    img: 'https://res.cloudinary.com/initial-cloud/image/upload/v1620870405/shirt_kbryht.png',
-    nombre: 'polera',
-    talla: 'L',
-    estado: "disponible",
-    color: "gris",
-    precio: "9990",
-    tipo: "Hombre",
-    categoria: "hombres"
-  },
-  {
-    id: 2,
-    img: 'https://res.cloudinary.com/initial-cloud/image/upload/v1620870405/shirt_kbryht.png',
-    nombre: 'pantalon',
-    talla: 'S',
-    estado: "disponible",
-    color: "negro",
-    precio: "5990",
-    tipo: "Hombre",
-    categoria: "hombres"
-  },
-  {
-    id: 3,
-    img: 'https://res.cloudinary.com/initial-cloud/image/upload/v1620870405/shirt_kbryht.png',
-    nombre: 'polera',
-    talla: 'L',
-    estado: "disponible",
-    color: "gris",
-    precio: "9990",
-    tipo: "Hombre",
-    categoria: "hombres"
-  },
-  {
-    id: 4,
-    img: 'https://res.cloudinary.com/initial-cloud/image/upload/v1620870405/shirt_kbryht.png',
-    nombre: 'polera',
-    talla: 'L',
-    estado: "disponible",
-    color: "gris",
-    precio: "9990",
-    tipo: "Hombre",
-    categoria: "hombres"
-  },
-  {
-    id: 5,
-    img: 'https://res.cloudinary.com/initial-cloud/image/upload/v1620870405/shirt_kbryht.png',
-    nombre: 'polera',
-    talla: 'L',
-    estado: "disponible",
-    color: "gris",
-    precio: "9990",
-    tipo: "Hombre",
-    categoria: "hombres"
-  },
-  {
-    id: 6,
-    img: 'https://res.cloudinary.com/initial-cloud/image/upload/v1620870405/shirt_kbryht.png',
-    nombre: 'polera',
-    talla: 'L',
-    estado: "disponible",
-    color: "gris",
-    precio: "9990",
-    tipo: "Hombre",
-    categoria: "hombres"
-  },
-  {
-    id: 7,
-    img: 'https://res.cloudinary.com/initial-cloud/image/upload/v1620870405/shirt_kbryht.png',
-    nombre: 'polera',
-    talla: 'L',
-    estado: "disponible",
-    color: "gris",
-    precio: "9990",
-    tipo: "Hombre",
-    categoria: "hombres"
-  },
-  {
-    id: 8,
-    img: 'https://res.cloudinary.com/initial-cloud/image/upload/v1620870405/shirt_kbryht.png',
-    nombre: 'polera',
-    talla: 'L',
-    estado: "disponible",
-    color: "gris",
-    precio: "9990",
-    tipo: "Hombre",
-    categoria: "hombres"
-  },
-
-]
-
-export const Ofertas = () => {
-  return (
-    <div>
-      <Header />
-
-      
-        <section className="container-titulo-ofertas">
-          <h2> OFERTAS </h2>
-        </section>
-      
-
-      <section className="ofertas-imagenes">
-        <div className="ofertas-ropa-ordenada" >
-          {
-            productos.map((producto) => {
-              return <Card key={producto.id} imagen={producto.img} />
-            })
-          }
-        </div>
-      </section>
-
-      <section className="ofertas-soporte">
-        <Soporte />
-      </section>
-
-    </div>
-
-  )
-}
