+++ src/views/Tipo.js
@@ -1,0 +1,12 @@
+import { Header } from "../components/Header"
+import { DetalleTipo } from "../components/DetalleTipo"
+
+
+export const Tipo = () => {
+  return (
+    <div>
+      <Header />
+      <DetalleTipo />
+    </div>
+  )
+}
